requirement.txt
PyQt5
paho-mqtt
sqlite3
pyqt5-tool